core: vmap gq format viz selected text alias of gw
AstroNvim disables the gq command in favor of the "more modern" gw
command. This commit restores the ability to use the gq command, but
treats it like an alias to gw

gw{motion}
Format the lines that {motion} moves over.

Similar to gq but puts the cursorback at the same position in the text.
However, 'formatprg' and 'formatexpr'are not used.

{Visual}gw
Format the highlighted text as with "gw"