refactor: rename stock plugin files, .template.lua
For files that were generated with the AstroNvim template located at:

https://github.com/AstroNvim/template

Replace their default .lua extension and append .template.lua to signify
that the file was originally included when cloning the template.

This isn't strictly necessary, it's just a convention for me to keep
track of what came with the clone initially and what I've added custom
to the ./lua/plugins folder.